pyproject.toml
[project]
name = "q-learning-homework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "box2d>=2.3.10",
    "gymnasium>=1.0.0",
    "matplotlib>=3.10.0",
    "np>=1.0.2",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pygame>=2.6.1",
    "pyqt6>=6.8.1",
    "ruff>=0.9.7",
    "seaborn>=0.13.2",
    "stable-baselines3>=2.5.0",
    "swig>=4.3.0",
    "tensordict",
    "tensordict-nightly>=2025.2.3",
    "torch>=2.6.0",
    "torchrl-nightly>=2025.2.3",
    "tqdm>=4.67.1",
]

[tool.uv.sources]
tensordict = { git = "https://github.com/pytorch-labs/tensordict" }
